Remove subcrates `CHANGELOG`
At this stage, it is important to allow the library to change rapidly.
Because of this, keeping a log of changes can be counter-productive. We
do not want pin down design decisions by writing detailed changelogs
(sunk cost fallacy).

Once the library and its different parts start becoming more stable and
mature, we will reintroduce changelogs accordingly.

For now, we will keep a main `CHANGELOG` file just for the `iced` crate.