Add support for setting window icon
This adds a new property from Settings::window::iconand a Icon struct which can be converted to winit::window::Icon.

It also adds code to display this icon in Application::run. Due to the fact that the Icon struct is non copyable, I also had to remove the Copy trait from all Settings, both in `iced` and `iced_winit`.