POC: layout can access widget dimensions via state
Only implemented for the button widget, but possible for all
widgets. The button width/height are stored in the widget state,
and that state is now a referance in the layout, meaning that
during the diff state, the state can be `diff`ed between the cache
(previous loop) and the requested values. Then addind in the change
in time will allow iced to calculate where the animation should be
for the current render.