Add method for UserInterface::build to request winit redraw
Using ControlFlow::WaitUntil lets us ask winit to redraw at a
time in the future. The previous commit makes `diff_mut` return
the most recent time a widget needs to redraw. We needed a way to
pass that information to the UserInterface::run closure and thus
winit. This does that.