Revert widget dimensions being stored as an animation
In a previous version of the animation RFC I thought it would make
sense for each widget to store their dimensions as a `Animation`
rather which was a wrapper around `Length`. This meant that each
widget would have to have an animate_width + animate_height methods
which isn't desierable. Ive decided I would prefer these to be
handled by a parent container widget. It seems more reasonable to
have each widget just Length::Fill in the animation axis, and have
the parent container animate. This reduces code dupication, and
allows for cimplex animations to be calculated in one location.