Use Animation API to build timeline
The timeline lets animations:
- chain together
- animate values in parallel, but at different rates
- loop more complicated animations.
- extending or modifying animations while they are running

This API sets the timeline based on Duration. But it should be
possible to expand this to use "speed" based animations. This
would be useful for animations that we don't know how far they
will move at runtime before reversing, i.e. a widget growing on
hover. The widget may only grow to 50% of the size before the user
unhovers the widget. Meaning that time based the widget would
shrink at a different rate than it grew. A speed based animation
would make the animation grow and shrink at the same rate,
reguardless of starting and ending position.