Use keyframes as a trait rather than global struct
Rather than having a type to hold data for all widgets,
each animatable widget must implement `Keyframe` on its own
`Keyframe` struct.

This allows for each widget to be able to reuse the animation
calculations even on custom widgets without having to fork iced.

This is more efficient as well because the each animation doesn't
have to "carry" the unused data that different widgets use. The
Playhead is also now lazily evaluated. This isn't because of the
triat implementation, but required to make it work, which happens
to also be more efficient.

This implementation isn't done yet, but it compiles.