Calculate lower & upper bounds for keyframe trait
This is necessary to know what percentage of the animation is done
for each step of the animation. This lets us calculate the current
value.

This was done before in da02689cdb8afc0179ae10c61d147870902b2519,
but this now works for the `Keyframe` trait that lets each widget
have its own implementation.