Track animation in windget state
Rather than needing to mutate any value, we can keep track of each
widget's start time and animation in the widget state. By using
different types of the tree::State we can guarentee that the
correct data is proveded for the animation and in the correct way.

A hash is generated for each animation, this means that we don't
have to store the animation itself between renders, just tracking
if it changes, and all data stays on the stack and nice and fast.

We know when the hash changes that a new animation has been
provided and we can start the new animation from the beginning.

We track the start time of each animation across renders as we can
take the current time of each render from the start time of the
animation and that gives us a "playhead" through the animation,
which then can let us calculate the animation's values.

Those calculated values can then be set for the widget in the
interp step. This gives the effect of an animation with the dev
only having to provide a start and an end (an points inbetween).

TODO: The `row` widget still needs to take the newly calulated
values and apply them to itself so the animation can play.