native/subscription/unfold: terminate stream when user returns None
`futures::stream::unfold` builds a `Stream` from the provided seed
of type `T` and a `Future` returned by the `f` function. The resulting
`Stream` produces values as long the `Future` returns `Some` tuple
`(Item, T)`. If the `Future` returns `None`, `unfold` is supposed to
stop producing items.

In `iced`, the `subscription::unfold` function is based on the
`stream::unfold`. However the `Output` of the `Future` is defined as
`(Option<Message>, T)`. The `Future` can return a `None` `Message`
and the `Stream` will keep running.

The above behaviour is exemplified in `examples/download_progress`.
When the download is finished (either because it is complete or
if an error occurred), the `Future` is still being called. Adding
a `println` statement shows that the `Future` is called repeatedly.
The original code added a `pending().await` as a workaround. This
means that the executor `Tasks` is kept in queues even if it will
no longer make progress.

One way to solve this would be to keep the existing signature for
`subscription::unfold` and to `map` the `Future` output in such a
way that it return `None` if the provided `Option<Message>` is
`None`. However, this would be inconsistent with the `unfold`
function from the `futures` crate and it wouldn't be suitable
since the `state` is no longer usable after the `Stream`
completes.

This commit aligns the `subscription::unfold` function with its
`futures` counterpart. In the `download_progress` example, after
a download is finished, the `Task` in no longer kept in queues.

This changes a bit the implementation for the `websocket` example.
Previous implementation returned from the `Future` with a `None`
message and an unchanged state when the websocket or input are
awaken but the result doesn't need to be propagated. With this
commit, in such cases the implementation loops on the wbesocket &
input async select awaiting for the next message to be propagated.