Use `glutin-winit` in `integration_opengl` example
Glutin 0.30 changes the API to not depend on winit, and instead take a
raw window handle. `glutin-winit` is a helper to make it a bit easier to
create a GL context on top of winit.

Things like `iced_winit::conversion::window_event` are problematic here
since we end up using a different winit version from `iced_winit`, and
that function is quite complicated to duplicate. But we don't need all
of it.

Duplicating isn't so unreasonable. It's probably expected that
integrations of Iced could not rely on conversions from `iced_winit`,
since likely use something other than winit to get a window and OpenGL
context, and in any case are not using the Iced fork of `winit`.