Add env var to default to software renderer
This allows a user for whom wgpu initializes but doesn't work properly
to switch to the software renderer without having to re-compile with
different features.

This could be done in a more complicated way with a `;` separated list
of renderer backends in the order to try them, but there's no need for
that when there are only two options.