Do not require `Copy` for text `StyleSheet::Style`
For most widgets, `Style` only requires `Default`. A few require
`Clone`. Only this one requires `Copy`.

Some of the types in the default theme has a custom variant requiring
`Box<dyn Trait>`, or `Rc<dyn Trait>` to provide `Clone`, but this isn't
possible if `Copy` is required.

It would be good to also address the inconsistency of requiring `Clone`
in some places and not others.

This removes `style/src/text.rs` which is unused in this branch and thus
confusing. If there's a reason to keep it, that can be removed from the
change.