Add Layout operation to `Container`
An operation that will return a message to the application with
the dimensions of the widget as a `Rectangle`.

`container::get_layout(CONTAINER_ID.clone(), Message::GotLayout)`

The only slightly unusual think about this operation is that
creating the operation requires the return message. This API
is the same as `Command::Perform`, and operations are just
`Commands`, so I believe this should be fine.