add platform specific settings for x11 window type
We use a couple of target_os configuration predicates to make sure we
are on a unix system. These were used by winit up to version "0.27" and
they still imply the "free_unix" configuration predicate that winit uses
since version "0.28".

More x11 or wayland platform specific settings can be added in the future.