feat: buttons hover and pressed animations
Before this commit, hovering a button "deviates" the color (it darkens
light ones and vice-versa) and pressing it reverts back to the original
color. With animations, it makes more sense to deviate the base color
one time on hover, and deviate it a second time on press, so that going
back from pressed to just active just goes from light/dark to dark/light
instead of dark/light to light/dark then back to dark/light.