Move `status` of button from widget to `State`
When the button is being recreated as part of an overlay
like in an `iced_aw::ContextMenu`, but the state will be kept
then the button will be rendered with the wrong status
(the default is always `None`,
so it will fall back to `Status::Disabled` rendering).

By moving the `status` from the widget data to the persistent `State`
of the widget tree, the rendering works as expected.